Added formatting for responses gathered from examination, and began work on if functions to determine if zip code test will be run.
Added functions to format answers given by user, then provided a summary to the user, then started working on the if functions that will be used to recommend the user gets tested then run our zip code test. Also added an explaination statement to our zip code test.  Tried to import the zip code test python file at the beginning of the program, but found this caused the examination to be bypassed. Will attempt to debug this issue, as it is cumbersome to have to import the same python file for every instance it should be run for.