Uploaded PPT Slide for Sprint 3 and fixed syntax errors in main code.
Uploaded the PPT slide for Project sprint 3. Also fixed various syntax errors in the main code, such as the misspelling of the word 'vomiting' and an improperly spelled variable name. Also changed the number of symptoms required for the under 4 symptoms ifs to trigger, as I wanted 4 and under for that.